pyproject.toml
[project]
name = "project1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "numpy>=2.4.1",
    "pillow>=12.1.0",
    "requests>=2.32.5",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "onnxruntime>=1.20.0",
    "uvicorn>=0.40.0",
]
